新增文件上传功能

+++ spring-platform-base/spring-platform-model/src/main/java/com/platform/model/entity/basic/SysFile.java
@@ -38,7 +38,7 @@ public class SysFile extends BaseEntity {
     /**
      * 文件大小
      */
-    private Double fileSize;
+    private Long fileSize;
 
     /**
      * 文件Hash值
